backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-django==4.7.0
flake8==6.1.0
black==23.11.0
django-filter==23.5



